+++ src/test/java/oneAppTests/SceneTests/SceneTests.java
@@ -1,0 +1,94 @@
+package oneAppTests.SceneTests;
+
+import io.appium.java_client.android.AndroidDriver;
+import oneAppTests.BaseAppiumTest;
+import oneAppTests.TestCaseSetup;
+import org.junit.Before;
+import org.junit.Test;
+import page.activity.PrimaryActivity;
+import page.fragment.*;
+
+import static org.junit.Assert.assertTrue;
+
+public class SceneTests extends BaseAppiumTest{
+
+    private static final String TEST_NAME = SceneTests.class.getName();
+
+    private static AndroidDriver driver = TestCaseSetup.getDriver();
+    private String sceneName = "Test Scene";
+
+    private DashboardFragmentPage dashboardFragmentPage;
+    private PrimaryActivity primaryActivity;
+    private DevicesFragmentPage devicesFragmentPage;
+    private AddSceneFragmentPage addSceneFragmentPage;
+    private AddActionsFragmentPage addActionsFragmentPage;
+    private SceneInformationFragmentPage sceneInformationFragmentPage;
+
+    @Override
+    public String getTestName() {
+        return TEST_NAME;
+    }
+
+    @Before
+    public void setUp() {
+        primaryActivity = new PrimaryActivity(driver);
+        dashboardFragmentPage = new DashboardFragmentPage(driver);
+        devicesFragmentPage = new DevicesFragmentPage(driver);
+        addSceneFragmentPage = new AddSceneFragmentPage(driver);
+        addActionsFragmentPage = new AddActionsFragmentPage(driver);
+        sceneInformationFragmentPage = new SceneInformationFragmentPage(driver);
+    }
+
+    @Test
+    public void sceneCreateTest() throws InterruptedException {
+        primaryActivity.devicesButton.click();
+        Thread.sleep(4000L);
+
+        devicesFragmentPage.moreOptionsButton.click();
+        devicesFragmentPage.getDevicesFragmentPage();
+        devicesFragmentPage.addSceneButton.click();
+
+        addSceneFragmentPage.sceneNameTextBox.sendKeys(sceneName);
+        addSceneFragmentPage.addActionsButton.click();
+
+        addActionsFragmentPage.outletButton.click();
+        addActionsFragmentPage.onRadioButton.click();
+        addActionsFragmentPage.doneButton.click();
+        addActionsFragmentPage.saveButton.click();
+        assertTrue(devicesFragmentPage.createdScene.isDisplayed());
+    }
+
+    @Test
+    public void sceneDeleteTest() throws InterruptedException {
+        primaryActivity.dashboardButton.click();
+        Thread.sleep(6000L);
+
+        dashboardFragmentPage.sceneInfoButton.click();
+        Thread.sleep(4000L);
+
+        sceneInformationFragmentPage.editButton.click();
+
+        addSceneFragmentPage.deleteButton.click();
+        addSceneFragmentPage.confirmDeleteButton.click();
+    }
+
+    @Test
+    public void sceneEditTest() throws InterruptedException {
+        primaryActivity.dashboardButton.click();
+        Thread.sleep(6000L);
+
+        dashboardFragmentPage.sceneInfoButton.click();
+        Thread.sleep(4000L);
+
+        sceneInformationFragmentPage.editButton.click();
+
+        addSceneFragmentPage.removeButton.click();
+        addSceneFragmentPage.addActionsButton.click();
+
+        addActionsFragmentPage.outletButton.click();
+        addActionsFragmentPage.onRadioButton.click();
+        addActionsFragmentPage.doneButton.click();
+        addActionsFragmentPage.saveButton.click();
+        assertTrue(dashboardFragmentPage.sceneInfoButton.isDisplayed());
+    }
+}
